+++ Classes-IPhone/GHUnitIPhoneAppDelegate.h
@@ -19,7 +19,5 @@
 	GHUnitIPhoneViewController *viewController_;
 }
 
-@property (nonatomic, retain) IBOutlet UIWindow *window;
-
 @end