+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# Release TBD (0.3.1)
+# Release TBD (0.3.2)
+
+## 2008-04-05
+- Building as static library for iPhone
+- Adding in support for running single test case or test
+
+# Release 0.3.1
 
 ## 2008-03-22
 - Renamed TEST_CLI to GHUNIT_CLI